+++ src/hooks/useValidators.ts
@@ -9,6 +9,9 @@ export type ValidatorError =
   | 'tooShort'
   | 'outOfRangeString'
   | 'outOfRangeNumber'
+  | 'outOfRangeSelectionItems'
+  | 'tooFewSelectionItems'
+  | 'tooManySelectionItems'
 export type ValidatorResult = ValidatorError | undefined
 
 const notEmpty = (value: unknown): ValidatorResult => {
@@ -46,11 +49,23 @@ const lengthValidator = (value: string | undefined, bounds: [number | undefined,
   return mapping[boundsValidator(value?.length, bounds)]
 }
 
+const selectionValidator = (value: unknown[], bounds: [number | undefined, number | undefined]): ValidatorResult => {
+  const mapping: Record<BoundsValidatorResult, ValidatorResult> = {
+    range: 'outOfRangeSelectionItems',
+    lower: 'tooFewSelectionItems',
+    upper: 'tooManySelectionItems',
+    none: undefined
+  }
+  return mapping[boundsValidator(value?.length, bounds)]
+}
+
 const emailValidator = (value: string | undefined) => {
   if (!value || !validateEmail(value)) {
     return 'invalidEmail'
   }
 }
+
+
 
 type ValidatorTranslationType = Record<ValidatorError, string>
 
@@ -61,7 +76,10 @@ export const defaultValidatorTranslation: Translation<ValidatorTranslationType> 
     tooShort: 'The value requires at least {{min}} characters.',
     tooLong: 'The value requires less than {{max}} characters.',
     outOfRangeString: 'The value needs to have between {{min}} and {{max}} characters.',
-    outOfRangeNumber: 'The value must be between {{min}} and {{max}}.'
+    outOfRangeNumber: 'The value must be between {{min}} and {{max}}.',
+    outOfRangeSelectionItems: 'Between {{min}} and {{max}} items must be selected.',
+    tooFewSelectionItems: 'Select at least {{min}} items.',
+    tooManySelectionItems: 'Select at most {{max}} items.',
   },
   de: {
     notEmpty: 'Das Feld darf nicht leer sein.',
@@ -69,7 +87,10 @@ export const defaultValidatorTranslation: Translation<ValidatorTranslationType> 
     tooShort: 'Der Wert muss mindestens {{min}} Zeichen enthalten.',
     tooLong: 'Der Wert darf höchstens {{max}} Zeichen enthalten.',
     outOfRangeString: 'Der Wert muss zwischen {{min}} und {{max}} Zeichen lang sein.',
-    outOfRangeNumber: 'Der Wert muss zwischen {{min}} und {{max}} liegen.'
+    outOfRangeNumber: 'Der Wert muss zwischen {{min}} und {{max}} liegen.',
+    outOfRangeSelectionItems: 'Es müssen zwischen {{min}} und {{max}} Elemente ausgewählt werden.',
+    tooFewSelectionItems: 'Es müssen mindestens {{min}} Elemente ausgewählt werden.',
+    tooManySelectionItems: 'Es müssen maximal {{max}} Elemente ausgewählt werden.',
   },
 }
 
@@ -77,6 +98,7 @@ export const UseValidators = {
   notEmpty: notEmpty,
   length: lengthValidator,
   email: emailValidator,
+  selection: selectionValidator,
 }
 
 export const useTranslatedValidators = () => {
@@ -85,21 +107,28 @@ export const useTranslatedValidators = () => {
   return {
     notEmpty: (value: unknown) => {
       const result = notEmpty(value)
-      if(result) {
+      if (result) {
         return translation(result)
       }
     },
     length: (value: string | undefined, length: [number | undefined, number | undefined]) => {
       const [min, max] = length
       const result = lengthValidator(value, length)
-      if(result) {
-        return translation(result, { replacements: { min: min.toString(), max: max.toString() } } )
+      if (result) {
+        return translation(result, { replacements: { min: min.toString(), max: max.toString() } })
       }
     },
     email: (value: string) => {
       const result = emailValidator(value)
-      if(result) {
+      if (result) {
         return translation(result)
+      }
+    },
+    selection: (value: unknown[]| undefined, length: [number | undefined, number | undefined]) => {
+      const [min, max] = length
+      const result = selectionValidator(value, length)
+      if (result) {
+        return translation(result, { replacements: { min: min.toString(), max: max.toString() } })
       }
     }
   }